feat: move accessor to each version
