+++ swagger_marshmallow_codegen/codegen/v2/codegen.py
@@ -4,7 +4,6 @@ import logging
 from collections import namedtuple
 from functools import partial
 from prestring import PreString
-from prestring.python import Module
 from prestring.utils import LazyFormat, LazyKeywordsRepr
 from dictknife import deepequal, deepmerge
 from collections import defaultdict
@@ -14,28 +13,11 @@ from swagger_marshmallow_codegen.langhelpers import (
     titleize,
     clsname_from_path,
 )
-from .options import CodegenTargetDict
-
+from ..options import CodegenTargetDict
+from ..context import Context
 
 logger = logging.getLogger(__name__)
 NAME_MARKER = "x-marshmallow-name"
-
-
-class Context:
-    def __init__(self, m=None, im=None):
-        self.m = m or Module(import_unique=True)
-        self.im = im or self.m.submodule()
-
-    def from_(self, module, name):
-        logger.debug("      import: module=%s, name=%s", module, name)
-        self.im.from_(module, name)
-
-    def import_(self, module):
-        logger.debug("      import: module=%s", module)
-        self.im.import_(module)
-
-    def new_child(self):
-        return self.__class__(self.m.submodule(newline=False), self.im)
 
 
 class CodegenError(Exception):
