+++ swagger_marshmallow_codegen/codegen/v3/codegen.py
@@ -1,0 +1,1 @@
+from ..v2.codegen import Codegen
