+++ swagger_marshmallow_codegen/driver.py
@@ -1,7 +1,6 @@
 # -*- coding:utf-8 -*-
 import logging
 from . import loading
-from .accessor import Accessor
 from .resolver import Resolver
 from .codegen import Codegen, SchemaWriter
 from .dispatcher import FormatDispatcher
@@ -13,7 +12,6 @@ logger = logging.getLogger(__name__)
 class Driver:
     dispatcher_factory = FormatDispatcher
     resolver_factory = Resolver
-    accessor_factory = Accessor
     codegen_factory = Codegen
 
     def __init__(self, options):
@@ -37,8 +35,7 @@ class Driver:
     def create_codegen(self):
         dispatcher = self.dispatcher_factory()
         resolver = self.resolver_factory(dispatcher)
-        accessor = self.accessor_factory(resolver)
-        return self.codegen_factory(accessor)
+        return self.codegen_factory(resolver)
 
 
 class LegacyDriver(Driver):
